factor PairF out of PairGraph into own module
