+++ pate.cabal
@@ -168,6 +168,7 @@ library
                        What4.JSON,
                        Data.Parameterized.SetF,
                        Data.Parameterized.SetCtx,
+                       Data.Parameterized.PairF,
                        Pate.SimState,
                        Pate.Equivalence,
                        Pate.Equivalence.Condition,
